moved writing routines to output
